requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
openpyxl==3.1.5
httpx==0.28.1
jinja2==3.1.5
python-dotenv==1.0.1